+++ skbio/alignment/_tabular_msa.py
@@ -19,7 +19,7 @@ import scipy.stats
 
 from skbio._base import SkbioObject, MetadataMixin, PositionalMetadataMixin
 from skbio.sequence import Sequence
-from skbio.sequence._iupac_sequence import IUPACSequence
+from skbio.sequence._grammared_sequence import GrammaredSequence
 from skbio.util._decorator import experimental, classonlymethod, overrides
 from skbio.util._misc import resolve_key
 from skbio.alignment._indexing import TabularMSAILoc, TabularMSALoc
@@ -35,7 +35,7 @@ class TabularMSA(MetadataMixin, PositionalMetadataMixin, SkbioObject):
 
     Parameters
     ----------
-    sequences : iterable of IUPACSequence, TabularMSA
+    sequences : iterable of GrammaredSequence, TabularMSA
         Aligned sequences in the MSA. Sequences must all be the same type and
         length. For example, `sequences` could be an iterable of ``DNA``,
         ``RNA``, or ``Protein`` sequences. If `sequences` is a ``TabularMSA``,
@@ -288,12 +288,12 @@ class TabularMSA(MetadataMixin, PositionalMetadataMixin, SkbioObject):
 
         Returns
         -------
-        TabularMSA, IUPACSequence, Sequence
+        TabularMSA, GrammaredSequence, Sequence
             A ``TabularMSA`` is returned when `seq_idx` and `pos_idx` are
-            non-scalars. A ``IUPACSequence`` of type ``msa.dtype`` is returned
-            when `seq_idx` is a scalar (this object will match the dtype of the
-            MSA). A ``Sequence`` is returned when `seq_idx` is non-scalar and
-            `pos_idx` is scalar.
+            non-scalars. A ``GrammaredSequence`` of type ``msa.dtype`` is
+            returned when `seq_idx` is a scalar (this object will match the
+            dtype of the MSA). A ``Sequence`` is returned when `seq_idx` is
+            non-scalar and `pos_idx` is scalar.
 
         See Also
         --------
@@ -554,12 +554,12 @@ class TabularMSA(MetadataMixin, PositionalMetadataMixin, SkbioObject):
 
         Returns
         -------
-        TabularMSA, IUPACSequence, Sequence
+        TabularMSA, GrammaredSequence, Sequence
             A ``TabularMSA`` is returned when `seq_idx` and `pos_idx` are
-            non-scalars. A ``IUPACSequence`` of type ``msa.dtype`` is returned
-            when `seq_idx` is a scalar (this object will match the dtype of the
-            MSA). A ``Sequence`` is returned when `seq_idx` is non-scalar and
-            `pos_idx` is scalar.
+            non-scalars. A ``GrammaredSequence`` of type ``msa.dtype`` is
+            returned when `seq_idx` is a scalar (this object will match the
+            dtype of the MSA). A ``Sequence`` is returned when `seq_idx` is
+            non-scalar and `pos_idx` is scalar.
 
         See Also
         --------
@@ -713,8 +713,8 @@ class TabularMSA(MetadataMixin, PositionalMetadataMixin, SkbioObject):
         Parameters
         ----------
         dictionary : dict
-            Dictionary mapping keys to ``IUPACSequence`` sequence objects. The
-            ``TabularMSA`` object will have its index labels set
+            Dictionary mapping keys to ``GrammaredSequence`` sequence objects.
+            The ``TabularMSA`` object will have its index labels set
             to the keys in the dictionary.
 
         Returns
@@ -924,7 +924,7 @@ class TabularMSA(MetadataMixin, PositionalMetadataMixin, SkbioObject):
 
         Yields
         ------
-        IUPACSequence
+        GrammaredSequence
             Each sequence in the order they are stored in the MSA.
 
         Examples
@@ -945,7 +945,7 @@ class TabularMSA(MetadataMixin, PositionalMetadataMixin, SkbioObject):
 
         Yields
         ------
-        IUPACSequence
+        GrammaredSequence
             Each sequence in reverse order from how they are stored in the MSA.
 
         Examples
@@ -1710,7 +1710,7 @@ class TabularMSA(MetadataMixin, PositionalMetadataMixin, SkbioObject):
 
         Parameters
         ----------
-        sequence : IUPACSequence
+        sequence : GrammaredSequence
             Sequence to be appended. Must match the dtype of the MSA and the
             number of positions in the MSA.
         minter : callable or metadata key, optional
@@ -1730,7 +1730,7 @@ class TabularMSA(MetadataMixin, PositionalMetadataMixin, SkbioObject):
             If neither `minter` nor `index` are provided and the MSA has a
             non-default index.
         TypeError
-            If the sequence object isn't an ``IUPACSequence``.
+            If the sequence object isn't a ``GrammaredSequence``.
         TypeError
             If the type of the sequence does not match the dtype of the MSA.
         ValueError
@@ -1792,7 +1792,7 @@ class TabularMSA(MetadataMixin, PositionalMetadataMixin, SkbioObject):
 
         Parameters
         ----------
-        sequences : iterable of IUPACSequence
+        sequences : iterable of GrammaredSequence
             Sequences to be appended. Must match the dtype of the MSA and the
             number of positions in the MSA.
         minter : callable or metadata key, optional
@@ -1816,7 +1816,8 @@ class TabularMSA(MetadataMixin, PositionalMetadataMixin, SkbioObject):
         ValueError
             If `index` is not the same length as `sequences`.
         TypeError
-            If `sequences` contains an object that isn't an ``IUPACSequence``.
+            If `sequences` contains an object that isn't a
+            ``GrammaredSequence``.
         TypeError
             If `sequence` contains a type that does not match the dtype of the
             MSA.
@@ -1910,10 +1911,10 @@ class TabularMSA(MetadataMixin, PositionalMetadataMixin, SkbioObject):
         else:
             sequence = sequences[0]
             expected_dtype = type(sequence)
-            if not issubclass(expected_dtype, IUPACSequence):
+            if not issubclass(expected_dtype, GrammaredSequence):
                 raise TypeError(
                     "Each sequence must be of type %r, not type %r"
-                    % (IUPACSequence.__name__, expected_dtype.__name__))
+                    % (GrammaredSequence.__name__, expected_dtype.__name__))
             expected_length = len(sequence)
 
         for sequence in sequences: