+++ skbio/io/format/phylip.py
@@ -126,10 +126,10 @@ Format Parameters
 -----------------
 The only supported format parameter is ``constructor``, which specifies the
 type of in-memory sequence object to read each aligned sequence into. This must
-be a subclass of ``IUPACSequence`` (e.g., ``DNA``, ``RNA``, ``Protein``) and is
-a required format parameter. For example, if you know that the PHYLIP file
-you're reading contains DNA sequences, you would pass ``constructor=DNA`` to
-the reader call.
+be a subclass of ``GrammaredSequence`` (e.g., ``DNA``, ``RNA``, ``Protein``)
+and is a required format parameter. For example, if you know that the PHYLIP
+file you're reading contains DNA sequences, you would pass ``constructor=DNA``
+to the reader call.
 
 Examples
 --------