+++ skbio/sequence/tests/test_rna.py
@@ -7,6 +7,7 @@
 # ----------------------------------------------------------------------------
 
 from __future__ import absolute_import, division, print_function
+import six
 
 import unittest
 
@@ -40,6 +41,10 @@ class TestRNA(unittest.TestCase):
         self.assertEqual(seq.reverse_transcribe(), DNA('ATAT'))
         self.assertEqual(seq, RNA('AUAU'))
 
+    def test_cannot_subclass(self):
+        with six.assertRaisesRegex(self, TypeError, "Subclassing disabled"):
+            class CustomSequence(RNA):
+                pass
 
 if __name__ == '__main__':
     unittest.main()